projects/capstone/requirements.txt
alembic==1.3.0
aniso8601==6.0.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
awscliv2==1.5.2
Babel==2.9.0
backcall==0.2.0
botocore==1.20.4
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
DateTime==4.3
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-environ==0.4.5
django-heroku==0.3.1
django-psycopg2-extension==0.0.5
docutils==0.15.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.9
Flask-Migrate==2.5.2
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4 
iniconfig==1.1.1 
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jwt==1.1.0
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
pipenv==2020.11.15
pluggy==0.13.1
postgres==3.0.0
prompt-toolkit==3.0.10
psycopg2==2.8.6
psycopg2-binary==2.8.2
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2019.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Searchable==1.2.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
toml==0.10.2 
urllib3==1.26.3
validators==0.18.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
waitress==1.4.4
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
zope.interface==5.2.0
